% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_imputation.R
\name{ALS_BS_missing_impute}
\alias{ALS_BS_missing_impute}
\title{Missing Imputation Based on Multiple Augmented Reduced Rank Regression via Soft-thresholding}
\usage{
ALS_BS_missing_impute(
  X_tot_incom,
  Y_org_list,
  n_mod_B,
  B_s_list,
  UB_s_list,
  VB_s_list,
  modules_B,
  n_mod_S = n_mod_B,
  S_s_list,
  U_s_list,
  V_s_list,
  modules_S = modules_B,
  modules_index_S,
  n_sample,
  lambdaBs,
  lambdaSs,
  bound,
  max_iter,
  Binvolved,
  Sinvolved,
  orth_sol,
  loss_comp,
  col_index
)
}
\arguments{
\item{X_tot_incom}{A matrix (#features x #samples), the concatenated version of
outcome matrices with NAs, e.g. \link{X1,X2,X3} if there are three cohorts in total.}

\item{Y_org_list}{A list of matrices, covariate matrices for each cohort.}

\item{n_mod_B}{A numeric, number of covariate-related modules included.}

\item{B_s_list}{A list of matrices, covariate effects of all modules. Initial values for optimization.}

\item{UB_s_list}{A list of matrices, loadings for covariate effects of all modules. Initial values for optimization.}

\item{VB_s_list}{A list of matrices, scores for covariate effects of all modules. Initial values for optimization.}

\item{modules_B}{A list of numeric vectors, indicating which cohort has a covariate effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module.
e.g. c(1,1) means that it is a joint covariate effect of both Y1 and Y2;
c(0,1) means that it is an individual covariate effect of Y2.}

\item{n_mod_S}{A numeric, number of covariate-unrelated modules included. Initial values for optimization.}

\item{S_s_list}{A list of matrices, auxiliary structures of all modules. Initial values for optimization.}

\item{U_s_list}{A list of matrices, loadings for auxiliary structures of all modules. Initial values for optimization.}

\item{V_s_list}{A list of matrices, scores for auxiliary structures of all modules. Initial values for optimization.}

\item{modules_S}{A list of numeric vectors, indicating which cohort has an auxiliary effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module.
e.g. c(1,1) means that it is a joint auxiliary structure \link{S1,S2};
c(0,1) means that it is an individual auxiliary structure S2.}

\item{n_sample}{A numeric vector, number of samples for each cohort.}

\item{lambdaBs}{A numeric vector, penalties for each B module.}

\item{bound}{A numeric, convergence criteria for squared norm difference of B/S.}

\item{max_iter}{A numeric, max number of epochs for optimization.}

\item{Binvolved}{A logical, whether to generate covaraite effects, i.e. B and Y.}

\item{Sinvolved}{A logical, whether to generate auxiliary variation structures, i.e. S.}

\item{orth_sol}{A logical, whether to orthogonalize covariate matrices before optimization
must set as "TRUE".}

\item{loss_comp}{A logical, whether to compute loss for each epoch.}

\item{col_index}{A list of numeric vectors, i th vector represents
the index for i th cohort in terms of Xtot.}

\item{lambdaSS}{A numeric vector, penalties for each S module.}
}
\value{
:
A list contains the following:
X_tot: A matrix, completed version of X_tot_incom.
diff_list: A numeric vector, sum of squared difference between the current
and previous imputation.
time: A numeric, total optimization time.
B_s_list: A list of matrices, estimated covariate effects of all modules.
S_s_list: A list of matrices, estimated auxiliary structures of all modules.
loss_list: A numeric vector, loss for each epoch.
}
\description{
Impute the missing values of multi-cohort data with the aid of another multi-cohort data based on
alternating least square and soft-thresholding. It requires semi-orthogonality of predictors.
Optimization based on algorithm 2 from maRRR paper.
}
\examples{
ALS_BS_missing_impute(X_tot,Y_org_list,n_mod_B,B_s_list,UB_s_list,
VB_s_list,modules_B,n_mod_S=n_mod_B,S_s_list,U_s_list,V_s_list,
modules_S=modules_B,n_sample,lambdaBs,lambdaSs,bound,max_iter,
Binvolved,Sinvolved,orth_sol,loss_comp,col_index)
}
\author{
Jiuzhou Wang
}
