% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maRRR.R
\name{ALS_UV}
\alias{ALS_UV}
\title{Multiple Augmented Reduced Rank Regression via Bilinear Decomposition}
\usage{
ALS_UV(
  X_tot,
  Y_org_list,
  n_mod_B,
  B_s_list,
  UB_s_list,
  VB_s_list,
  modules_B,
  n_mod_S = n_mod_B,
  S_s_list,
  U_s_list,
  V_s_list,
  modules_S = modules_B,
  n_sample,
  lambdaBs,
  lambdaSs,
  bound,
  max_iter,
  Binvolved,
  Sinvolved,
  orth_sol,
  loss_comp
)
}
\arguments{
\item{X_tot}{A matrix (#features x #samples), the concatenated version of outcome matrices.
e.g. \link{X1,X2,X3} if there are three cohorts in total.}

\item{Y_org_list}{A list of matrices, covariate matrices for each cohort.}

\item{n_mod_B}{A numeric, number of covariate-related modules included.}

\item{B_s_list}{A list of matrices, covariate effects of all modules. Initial values for optimization.}

\item{UB_s_list}{A list of matrices, loadings for covariate effects of all modules. Initial values for optimization.}

\item{VB_s_list}{A list of matrices, scores for covariate effects of all modules. Initial values for optimization.}

\item{modules_B}{A list of numeric vectors, indicating which cohort has a covariate effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module.
e.g. c(1,1) means that it is a joint covariate effect of both Y1 and Y2;
c(0,1) means that it is an individual covariate effect of Y2.}

\item{n_mod_S}{A numeric, number of covariate-unrelated modules included. Initial values for optimization.}

\item{S_s_list}{A list of matrices, auxiliary structures of all modules. Initial values for optimization.}

\item{U_s_list}{A list of matrices, loadings for auxiliary structures of all modules. Initial values for optimization.}

\item{V_s_list}{A list of matrices, scores for auxiliary structures of all modules. Initial values for optimization.}

\item{modules_S}{A list of numeric vectors, indicating which cohort has an auxiliary effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module.
e.g. c(1,1) means that it is a joint auxiliary structure \link{S1,S2};
c(0,1) means that it is an individual auxiliary structure S2.}

\item{n_sample}{A numeric vector, number of samples for each cohort.}

\item{lambdaBs}{A numeric vector, penalties for each B module.}

\item{bound}{A numeric, convergence criteria for squared norm difference of B/S.}

\item{max_iter}{A numeric, max number of epochs for optimization.}

\item{Binvolved}{A logical, whether to generate covaraite effects, i.e. B and Y.}

\item{Sinvolved}{A logical, whether to generate auxiliary variation structures, i.e. S.}

\item{orth_sol}{A logical, whether to orthogonalize covariate matrices before optimization.}

\item{loss_comp}{A logical, whether to compute loss for each epoch.}

\item{lambdaSS}{A numeric vector, penalties for each S module.}
}
\value{
: A list contains the following:

\item{time}{A numeric, total optimization time.}
\item{B_s_list}{A list of matrices, estimated covariate effects of all modules.}
\item{S_s_list}{A list of matrices, estimated auxiliary structures of all modules.}
\item{loss_list}{A numeric vector, loss for each epoch.}
}
\description{
Predict multi-cohort data from another multi-cohort data based on
alternating least square and blinear decomposition of estimates.
Optimization based on algorithm 1 from maRRR paper.
}
\examples{
ALS_UV(X_tot,Y_org_list,n_mod_B,B_s_list,UB_s_list,
VB_s_list,modules_B,n_mod_S=n_mod_B,S_s_list,U_s_list,V_s_list,
modules_S=modules_B,n_sample,lambdaBs,lambdaSs,bound,max_iter,
Binvolved,Sinvolved,orth_sol,loss_comp)
}
\author{
Jiuzhou Wang
}
