% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_gen}
\alias{data_gen}
\title{Generate ground truth B,S,X,Y for simulation use}
\usage{
data_gen(
  seed = 1,
  n_sample = c(100, 100),
  p_x = 100,
  p_y = 10,
  R_b_true = 1,
  R_s_true = 5,
  correlation = 0,
  orth_gen = FALSE,
  Binvolved = TRUE,
  Sinvolved = TRUE,
  modules_B = list(c(1, 1), c(1, 0), c(0, 1)),
  modules_index_B = list(c(1, 2), c(1), c(2)),
  sd_B = c(1, 1, 1),
  n_mod_B = length(modules_B),
  modules_S = modules_B,
  modules_index_S = modules_index_B,
  sd_S = sd_B,
  n_mod_S = length(modules_S)
)
}
\arguments{
\item{seed}{A numeric, random seed}

\item{n_sample}{A numeric vector, number of samples for each cohort}

\item{p_x}{A numeric, number of features in the observed data matrices}

\item{p_y}{A numeric, number of accompanying covariates}

\item{R_b_true}{A numeric, true rank for simulated B matrices}

\item{R_s_true}{A numeric, true rank for simulated S matrices}

\item{correlation}{A numeric, Pearson's correlation aiming to achieve in the simulated Y matrices}

\item{orth_gen}{A logical, whether to orthogonalize the simulated Y matrices}

\item{Binvolved}{A logical, whether to generate covaraite effects, i.e. B and Y}

\item{Sinvolved}{A logical, whether to generate auxiliary variation structures, i.e. S}

\item{modules_B}{A list of numeric vectors, indicating which cohort has a covariate effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module
e.g. c(1,1) means that it is a joint covariate effect of both Y1 and Y2;
c(0,1) means that it is an individual covariate effect of Y2}

\item{modules_index_B}{A list of numeric vectors, indicating the index of Yi included
e.g. c(1,2) will corresponds to c(1,1) in "modules_B";
e.g. c(1,3,5) will corresponds to c(1,0,1,0,1) in "modules_B"}

\item{sd_B}{A numeric vector, standard deviation of covariate coefficient matrices (B)}

\item{n_mod_B}{A numeric, number of covariate-related modules included}

\item{modules_S}{A list of numeric vectors, indicating which cohort has an auxiliary effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module
e.g. c(1,1) means that it is a joint auxiliary structure (S1,S2);
c(0,1) means that it is an individual auxiliary structure S2}

\item{modules_index_S}{A list of numeric vectors, indicating the index of Si included
e.g. c(1,2) will corresponds to c(1,1) in "modules_S";
e.g. c(1,3,5) will corresponds to c(1,0,1,0,1) in "modules_S"}

\item{sd_S}{A numeric vector, standard deviation of auxiliary structure matrices (S)}

\item{n_mod_S}{A numeric, number of covariate-unrelated modules included}
}
\value{
:
A list of a outcome matrix, covariate matices, covariate effects and auxiliary effects
\item{X_tot}{A matrix (#features x #samples), the concatenated version of outcome matrices
e.g. (X1,X2,X3) if there are three cohorts in total}
\item{Y_org_list}{A list of matrices, covariate matrices for each cohort}
\item{B_list}{A list of matrices, covariate effects of all modules}
\item{S_list}{A list of matrices, auxiliary structures of all modules}
}
\description{
Generate ground truth B,S,X,Y for simulation use
}
