% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{ini_gen}
\alias{ini_gen}
\title{Initialize starting values for optimization}
\usage{
ini_gen(
  seed = 1,
  n_sample = c(100, 100),
  p_x = 100,
  p_y = 10,
  R_b = 10,
  R_s = 10,
  n_mod_B = 3,
  n_mod_S = n_mod_B,
  Binvolved = TRUE,
  Sinvolved = TRUE
)
}
\arguments{
\item{seed}{A numeric, random seed}

\item{n_sample}{A numeric vector, number of samples for each cohort}

\item{p_x}{A numeric, number of features in the observed data matrices}

\item{p_y}{A numeric, number of accompanying covariates}

\item{R_b}{A numeric, upper bound for rank of simulated B matrices}

\item{R_s}{A numeric, upper bound for rank of simulated S matrices}

\item{n_mod_B}{A numeric, number of covariate-related modules included}

\item{n_mod_S}{A numeric, number of covariate-unrelated modules included}

\item{Binvolved}{A logical, whether to generate covaraite effects, i.e. B and Y}

\item{Sinvolved}{A logical, whether to generate auxiliary variation structures, i.e. S}
}
\value{
:
A list of initialized estimates.
\item{UB_s_list}{A list of matrices, loadings for covariate effects of all modules}
\item{VB_s_list}{A list of matrices, scores for covariate effects of all modules}
\item{B_s_list}{A list of matrices, covariate effects of all modules}
\item{U_s_list}{A list of matrices, loadings for auxiliary structures of all modules}
\item{V_s_list}{A list of matrices, scores for auxiliary structures of all modules}
\item{S_s_list}{A list of matrices, auxiliary structures of all modules}
}
\description{
Initialize starting values for optimization
}
