% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_penalty.R
\name{lambda_B_gen}
\alias{lambda_B_gen}
\title{Generate penalty for B}
\usage{
lambda_B_gen(B_list)
}
\arguments{
\item{B_list}{A list of matrices, covariate effects of all modules.}
}
\value{
all_lambda: A vector of penalties for all auxiliary structures
}
\description{
Generate best penalties for covariate effects B
based on random matrix theory from maRRR paper.
}
\examples{
lambdaBs = lambda_B_gen(B_list = dat[[3]])
}
\author{
Jiuzhou Wang
}
