% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_penalty.R
\name{lambda_S_gen}
\alias{lambda_S_gen}
\title{Generate penalty for S}
\usage{
lambda_S_gen(p_x, modules_index_S, n_sample)
}
\arguments{
\item{p_x}{A numeric, number of features of outcomes X.}

\item{modules_index_S}{list of numeric vectors, indicating the index of Si included
e.g. c(1,2) will corresponds to c(1,1) in "modules_S";
e.g. c(1,3,5) will corresponds to c(1,0,1,0,1) in "modules_S"}

\item{n_sample}{A numeric vector, number of samples for each cohort.}
}
\value{
all_lambda: A vector of penalties for all auxiliary structures
}
\description{
Generate best penalties for auxiliary structures S
based on random matrix theory from maRRR paper.
}
\examples{
lambdaSs = lambda_S_gen(p_x=100,
modules_index_S=list(c(1,2),c(1),c(2)),n_sample=c(100,100))
}
\author{
Jiuzhou Wang
}
