% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{loss_calculator}
\alias{loss_calculator}
\title{Calculate current}
\usage{
loss_calculator(
  X_tot,
  n_mod_B,
  n_mod_S,
  lambdaBs,
  lambdaSs,
  B_s_list,
  S_s_list,
  Y_list
)
}
\description{
It can be computationally heavy if high dimension.
}
