% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{mod_refine}
\alias{mod_refine}
\title{Concatenate all matrices in a module, including zeros to maintain the shape}
\usage{
mod_refine(X, mod, n_sample)
}
\arguments{
\item{X}{A matrix, full concatenated outcome matrix without zeros.}

\item{mod}{A numeric vector, indicating which cohort has a covariate effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module
e.g. c(1,1) means that it is a joint covariate effect of both Y1 and Y2;
c(0,1) means that it is an individual covariate effect of Y2}

\item{n_sample}{A numeric vector, number of samples for each cohort.}
}
\value{
temp: A matrix, concatenated outcome matrix for current module with zeros.
}
\description{
Generate multi-cohort X matrix from each cohort.
If a cohort is not included in the module, the corresponding entries will be zero.
}
