% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{thres}
\alias{thres}
\title{Replace all negative numbers with zeros in a vector}
\usage{
thres(vector)
}
\arguments{
\item{vector}{A vector}
}
\value{
new_vec: A vector without negative numbers
}
\description{
Replace all negative numbers with zeros in a vector
}
