% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{V_generator}
\alias{V_generator}
\title{Generate a random matrix or a zero matrix}
\usage{
V_generator(nc = 10, nr = 100, nonzero = 1)
}
\arguments{
\item{nc}{A numeric, number of columns.}

\item{nr}{A numeric, number of rows.}

\item{nonzero}{A numeric, only 1 or 0 to indicate whether nonzero is needed or not.}
}
\value{
V: A matrix.
}
\description{
A helper function. Generate V or V-shaped zero matrix.
}
