% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{Y_combiner_nonzero}
\alias{Y_combiner_nonzero}
\title{Concatenate predictors from multiple cohorts without zero-like matrix}
\usage{
Y_combiner_nonzero(Yi_list, mod)
}
\arguments{
\item{Yi_list}{A list of matrices, covariate matrices for each cohort}

\item{mod}{A numeric vector, indicating which cohort has a covariate effect in each module,
i th "1" in the j th vector means the i th cohort is included in the j th module
e.g. c(1,1) means that it is a joint covariate effect of both Y1 and Y2;
c(0,1) means that it is an individual covariate effect of Y2}
}
\value{
Y: A matrix.
}
\description{
Generate multi-cohort Y matrix for each module from each cohort.
If a cohort is not included in the module, there will be no corresponding entries.
}
